refactor: adding indexs
