+++ nuestra_senora_informativa/src/Components/index.ts
@@ -1,0 +1,8 @@
+export { default as Button } from './Button';
+export { default as ConfirmationModal } from './ConfirmationModal';
+export { default as CustomSelect } from './CustomSelect';
+export { default as Footer } from './Footer';
+export { default as Header } from './Header';
+export { default as InputForm } from './InputForm';
+export { default as LoadingSpinner } from './LoadingSpinner';
+export { default as ServiceCard } from './ServiceCard';
